+++ kyuubi-server/src/test/scala/org/apache/kyuubi/server/TempFileServiceSuite.scala
@@ -36,8 +36,8 @@ class TempFileServiceSuite extends WithKyuubiServer {
     .set(SERVER_TEMP_FILE_EXPIRE_TIME, Duration.ofMillis(expirationInMs).toMillis)
 
   test("file cleaned up after expiration") {
-    val tempFileService = KyuubiServer.kyuubiServer.backendService.sessionManager
-      .asInstanceOf[KyuubiSessionManager].tempFileService
+    val tempFileService =
+      server.backendService.sessionManager.asInstanceOf[KyuubiSessionManager].tempFileService
     (0 until 3).map { i =>
       val dir = Utils.createTempDir()
       writeToTempFile(new ByteArrayInputStream(s"$i".getBytes()), dir, s"$i.txt")
